connet-helper -> connect-helper rename
